update product, category, color, size

+++ backend/src/controller/category.js
@@ -1,4 +1,6 @@
 import Category from '../models/category.js';
+import Product from "../models/product.js"
+import mongoose from 'mongoose';
 
 export const getAll = async (req, res) => {
   try {
@@ -22,24 +24,30 @@ export const getAll = async (req, res) => {
 
 export const get = async (req, res) => {
   try {
+    
+    const {id} = req.params
+
+    // Kiểm tra trong mongoose nếu id không phải là một ObjectId thì trả về message
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(401).json({
+            message: "Không tìm thấy ID danh mục"
+        })
+    }
 
     // const categorys = await Category.findById(req.params.id).populate("products");
     const categorys = await Category.findById(req.params.id).populate("products");
-
-    if (!categorys) {
-      return res.json({
-        message: "Không tìm thấy Category!",
-      });
-    }4
-    const product = await Product.find({ categoryId: req.params.id })
+    if(!categorys){
+      return res.status(400).json({
+        message: "Không tồn tại danh mục bạn đang tìm"
+      })
+    }
     return res.status(200).json({
       message: "Lấy Category thành công!",
-      data: categorys,
-      product
+      ...categorys.toObject()
     });
   } catch (error) {
     return res.status(400).json({
-      message: error,
+      message: error.message,
     });
   }
 };
@@ -69,12 +77,22 @@ export const remove = async (req, res) => {
   try {
     const categoryId = req.params.id;
 
-    const categorys = await Category.findByIdAndDelete({ _id: req.params.id });
-    if (!categorys) {
-      return res.json({
-        message: "Không tìm thấy Category!",
-      });
+    const {id} = req.params
+
+    // Kiểm tra trong mongoose nếu id không phải là một ObjectId thì trả về message
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(401).json({
+            message: "Không tìm thấy ID danh mục"
+        })
     }
+
+    const categorys = await Category.findByIdAndDelete({ _id: req.params.id });
+    if(!categorys){
+      return res.status(400).json({
+        message: "Không tồn tại danh mục cần xóa"
+      })
+    }
+
     await Product.deleteMany({ categoryId });
     
     return res.status(200).json({
@@ -84,26 +102,38 @@ export const remove = async (req, res) => {
   }
   catch (error) {
     return res.status(500).json({
-      message: error,
+      message: error.message,
     })
   }
 }
 
 export const update = async (req, res) => {
   try {
-    const categorys = await Category.findByIdAndUpdate({ _id: req.params.id }, req.body, { new: true });
-    if (!categorys) {
-      return res.status(404).json({
-        message: "Không tìm thấy Category!",
-      });
+    
+    const {id} = req.params
+
+    // Kiểm tra trong mongoose nếu id không phải là một ObjectId thì trả về message
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(401).json({
+            message: "Không tìm thấy danh mục cần update"
+        })
     }
+
+    const categorys = await Category.findByIdAndUpdate({ _id: req.params.id }, req.body, { new: true });
+
+    if(!categorys){
+      return res.status(400).json({
+        message: "Không tồn tại danh mục cần cập nhật"
+      })
+    }
+
     return res.status(200).json({
       message: "Category đã được cập nhật thành công!",
       data: categorys,
     });
   } catch (error) {
     return res.status(500).json({
-      message: error,
+      message: error.message,
     });
   }
 };
