+++ backend/src/controller/size.js
@@ -1,4 +1,5 @@
 import Size from '../models/size.js';
+import mongoose from 'mongoose';
 
 export const getAll = async (req, res) => {
   try {
@@ -14,16 +15,20 @@ export const getAll = async (req, res) => {
     });
   } catch (error) {
     return res.status(400).json({
-      message: error,
+      message: error.message,
     });
   }
 };
 
 export const get = async (req, res) => {
   try {
-
+    const {id} = req.params
+    if(!mongoose.Types.ObjectId.isValid(id)){
+      return res.status(400).json({
+        message: "Không tìm thấy size trong database"
+      })
+    }
     const sizes = await Size.findById(req.params.id);
-
     if (!sizes) {
       return res.json({
         message: "Không tìm thấy size!",
@@ -35,7 +40,7 @@ export const get = async (req, res) => {
     });
   } catch (error) {
     return res.status(400).json({
-      message: error,
+      message: error.message,
     });
   }
 };
@@ -56,13 +61,20 @@ export const create = async (req, res) => {
     });
   } catch (error) {
     return res.status(400).json({
-      message: error,
+      message: error.message,
     });
   }
 };
 
 export const remove = async (req, res) => {
   try {
+    const {id} = req.params
+    if(!mongoose.Types.ObjectId.isValid(id)){
+      return res.status(400).json({
+        message: "Không tìm thấy size trong database"
+      })
+    }
+
     const sizes = await Size.findByIdAndDelete({ _id: req.params.id });
     if (!sizes) {
       return res.json({
@@ -76,13 +88,21 @@ export const remove = async (req, res) => {
   }
   catch (error) {
     return res.status(500).json({
-      message: error,
+      message: error.message,
     })
   }
 }
 
 export const update = async (req, res) => {
   try {
+
+    const {id} = req.params
+    if(!mongoose.Types.ObjectId.isValid(id)){
+      return res.status(400).json({
+        message: "Không tìm thấy size trong database"
+      })
+    }
+
     const sizes = await Size.findByIdAndUpdate({ _id: req.params.id }, req.body, { new: true });
     if (!sizes) {
       return res.status(404).json({
@@ -95,7 +115,7 @@ export const update = async (req, res) => {
     });
   } catch (error) {
     return res.status(500).json({
-      message: error,
+      message: error.message,
     });
   }
 };
